Exercise 01 corrections 2

+++ e01/E01.java
@@ -5,7 +5,7 @@ class E01 {
         float y = 5.5f;
         boolean flag = true;
         char letter = 'c';
-        short year = 2022;
+        int year = 2022;
         long population = 7000000000L;
         byte num = 100;
         double pi = 3.13457599923384753929348;
